backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ff==0.0.10
idna==3.11
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
reportlab==4.4.6
requests==2.32.5
sqlparse==0.5.4
stripe==14.1.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
